fix/database-model:fix the way models return when called
