+++ Entregas/BE-4/.gitignore
@@ -1,4 +1,4 @@
-
 node_modules
 .env
+package-lock.json
