--- Entregas/BE-4/src/db/models/address.js
@@ -1,44 +1,0 @@
-module.exports = (sequelize,DataTypes,User) => {
-    const Address = sequelize.define('Address', {
-        id: {
-            type: DataTypes.INTEGER,
-            primaryKey: true,
-            autoIncrement: true,
-        },
-        user_id: {
-            type: DataTypes.INTEGER,
-            allowNull: false,
-            references : {
-                model : User,
-                key: 'id'
-                }
-        },
-        state: {
-            type: DataTypes.STRING,
-            allowNull: false,
-        },
-        city: {
-            type: DataTypes.STRING,
-            allowNull: false,
-        },
-        neighborhood: {
-            type: DataTypes.STRING,
-            allowNull: false,
-        },
-        street: {
-            type: DataTypes.STRING,
-            allowNull: false,
-        },
-        number: {
-            type: DataTypes.STRING,
-            allowNull: false,
-        },
-        complement: {
-            type: DataTypes.STRING,
-            allowNull: true,
-        },
-    }, {
-        tableName: 'addresses',timestamps : false});
-
-    return Address;
-}
